Refactored code so that conversion from QImage to texture happens in moveface-image-source itself
